moved logic for displaying error into errorHandlers
